requirements.txt
psycopg2-binary==2.9.9
pyarrow==14.0.1
pyyaml==6.0.1
pandas==2.1.4
python-dateutil==2.8.2
pandera==0.18.0
sqlalchemy==2.0.23

# Testing dependencies
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1